Load current collection when routing to new /collections/{collection_id} page
This adds the new Pages.COLLECTION for rendering specific collections of
papers, updates loadCollections in HomeService to also set the current
collection based on `collection_id` route parameter, and fetches
metadata for the current collection's paperIds once the current
collection is set.

In the home-gallery component, this adds an "active" style for the
current collection nav item.

Follow-up commit will show the current collection's papers
(or history service papers if current collection is undefined, i.e.,
user is on the home page).