Highlight query text when Ask Lumi is clicked
Prior to this change, the text highlight disappears when Ask Lumi is triggered in the smart ask menu. After this change, the text will be highlighted in light green, while the user types a question and while the response loads. The highlight becomes the normal green once the response loads. (Clicking away from the menu clears the intermediate state highlight.)